set event width correctly on cleanUp
This change calculates the new width by using the width of the containing column and the left of the event. The fix is to treat .css() as always returning a computed size.

Previously, the cleanUp method calculated the "left" by using .css('left') and assuming the result was the percentage as seen in the DOM.

This produces reasonable looking results most of the time, but entirely by accident. The css() method returns computed sizes, so using it as a percentage produces mostly invalid results. This isn't perfect (really we need to reflow every event to make it totally awesome) but at least prevents "size zero" events for now.